==> analysis_script_sab/main_analysis/03.GershmanModellingMain_sab.rmd <==
---
title: "03.GershmanModellingMain"
author: "Sab Lam"
date: "18/05/2021"
output: html_document
---

Load the packages

```{r setup, include=FALSE, cache = FALSE}
require(knitr)
library(tidyverse)
library(dplyr)
library(tidyr)
library(rstan)
library(ggplot2)
library(loo)

source('src/functions.R')

knitr::opts_chunk$set(echo = TRUE)
```


Read in processed gershman data

```{r}
gershman_data_clean <- read.csv(file = "data/processed/gershman/data.csv")%>%
  mutate(trial_nr = trial_nr+1)%>% #somehow the gershman data has trial numbers 0-299, we're now changing the trial numbers to 1-300
  mutate_at(vars(starts_with('kalman')), scale) #z transformation/standardisation

gershman_data_clean %>% count(!is.na(total_trial_nr), prolific) 

# all people get 300 trials but 3 participants got unusual total_trial_nr (the total_trial_nr looks like participants are skipping every other trial no.)

######  check using the following script which are commented out  ###### 
# gershman_data_clean$correct_total_trial_nr = rep(1:300,547)
# gershman_data_clean$total_trial_check = gershman_data_clean$total_trial_nr - gershman_data_clean$correct_total_trial_nr
# gershman_data_clean <- gershman_data_clean%>%
#   filter(total_trial_check != 0)%>%
#   dplyr::distinct(prolific)

gershman_data_clean <- gershman_data_clean%>%
   mutate(total_trial_nr = rep(1:300,547)) 
```


Number of subjects (strictly positive int)

```{r}
nSubject <- gershman_data_clean%>%
  select(prolific)%>%
  dplyr::distinct(prolific)

N = as.numeric(nrow(nSubject))
```


Number of trials (strictly positive int)

```{r}
nTrial <- gershman_data_clean%>%
  select(total_trial_nr) 

T = as.numeric(max(nTrial))
```


Tsubj: Max number of trials per participants (1D array of ints)

```{r}
Tsubj <- gershman_data_clean%>%
  select(prolific, total_trial_nr, chose_right)

Tsubj <- Tsubj%>%
  pivot_wider(names_from = total_trial_nr, values_from = chose_right)

Tsubj <- Tsubj %>% remove_rownames %>% column_to_rownames(var = "prolific")

Tsubj$max_trial = rowSums(!is.na(Tsubj))

Tsubj <- Tsubj%>%
  select(max_trial)

Tsubj <- as.data.frame(t(Tsubj))

Tsubj <- t(Tsubj)

Tsubj = as.vector(Tsubj)
```


Kalman Value Difference (i.e. V) on that trial — (rows: participants, columns : trials)

```{r}
V <- gershman_data_clean%>%
  select(prolific, kalman_value_difference, total_trial_nr)

V <- V%>%
pivot_wider(names_from = total_trial_nr, values_from = kalman_value_difference) 
 
V <- V%>% remove_rownames %>% column_to_rownames(var = "prolific")

V <- as.data.frame(V)
```


Kalman Sigma Difference (i.e. RU) on that trial — (rows: participants, columns : trials)

```{r}
RU <- gershman_data_clean%>%
  select(prolific, kalman_sigma_difference, total_trial_nr)

RU <- RU%>%
pivot_wider(names_from = total_trial_nr, values_from = kalman_sigma_difference) 
 
RU <- RU%>% remove_rownames %>% column_to_rownames(var = "prolific")

RU <- as.data.frame(RU)
```


Kalman Value Difference scaled by total uncertainty (i.e. V/TU) on that trial — (rows: participants, columns : trials)

```{r}
VTU <- gershman_data_clean%>%
  select(prolific, kalman_weighted_value_difference, total_trial_nr,)

VTU<-VTU%>%
pivot_wider(names_from=total_trial_nr, values_from=kalman_weighted_value_difference) 


VTU<-VTU%>% remove_rownames %>% column_to_rownames(var="prolific")

VTU <- as.data.frame(VTU)
```


Choice: Array of ints containing the choice made for each trial and participant (i.e. option chosen out of 2 : 0 or 1) — (rows: participants, columns: trials)

```{r}
# changed code in Gershman model 1 from to "categorical" to "bernoulli" as choices are 0s or 1s

choice <- gershman_data_clean%>%
  select(prolific, chose_right, total_trial_nr)

choice <- choice%>%
pivot_wider(names_from = total_trial_nr, values_from = chose_right) 

choice<-choice%>% remove_rownames %>% column_to_rownames(var="prolific")

choice <- as.data.frame(choice)
```


Create a datalist of all data

```{r}
GershmandataList<-list(N=N,
               T=T,
               Tsubj=Tsubj,
               V=V,
               RU=RU,
               VTU=VTU,
               choice=choice)
```


(VB) Load the model into an object

```{r}
gershman_model1 <- stan_model("Models/ExploreTask/models/model_1/model_1.stan", verbose = TRUE)

gershman_model2 <- stan_model("Models/ExploreTask/models/model_2/model_2.stan", verbose = TRUE)

gershman_model3 <- stan_model("Models/ExploreTask/models/model_3/model_3.stan", verbose = TRUE)

gershman_model4 <- stan_model("Models/ExploreTask/models/model_4/model_4.stan", verbose = TRUE)

gershman_model5 <- stan_model("Models/ExploreTask/models/model_5/model_5.stan", verbose = TRUE)
```


(MCMC) Load the model into an object

```{r}
gershman_model1file <- ("Models/ExploreTask/models/model_1/model_1.stan")

gershman_model2file <- ("Models/ExploreTask/models/model_2/model_2.stan")

gershman_model3file <- ("Models/ExploreTask/models/model_3/model_3.stan")

gershman_model4file <- ("Models/ExploreTask/models/model_4/model_4.stan")

gershman_model5file <- ("Models/ExploreTask/models/model_5/model_5.stan")
```


(VB) Run model 1

```{r}
startTime = Sys.time(); print(startTime) 

fit_gershman1_vb<- vb(gershman_model1,
                data    = GershmandataList,
                seed    = 1450154626) 

endTime = Sys.time(); print(endTime) 
cat("It took",as.character.Date(endTime - startTime), "to fit Gershman Model 1\n")

#Summaries and plots for model 1
summary(fit_gershman1_vb)
```


(VB) Run model 2

```{r}
startTime = Sys.time(); print(startTime) 

fit_gershman2_vb<- vb(gershman_model2,
                data    = GershmandataList,
                seed    = 1450154626) 

endTime = Sys.time(); print(endTime) 
cat("It took",as.character.Date(endTime - startTime), "to fit Gershman Model 2\n")

#Summaries and plots for model 2
summary(fit_gershman1_vb)
```


(VB) Run model 3

```{r}
startTime = Sys.time(); print(startTime) 

fit_gershman3_vb<- vb(gershman_model3,
                data    = GershmandataList,
                seed    = 1450154626) 

endTime = Sys.time(); print(endTime) 
cat("It took",as.character.Date(endTime - startTime), "to fit Gershman Model 3\n")

#Summaries and plots for model 3
summary(fit_gershman3_vb)
```


(VB) Run model 4

```{r}
startTime = Sys.time(); print(startTime) 

fit_gershman4_vb<- vb(gershman_model4,
                data    = GershmandataList,
                seed    = 1450154626) 

endTime = Sys.time(); print(endTime) 
cat("It took",as.character.Date(endTime - startTime), "to fit Gershman Model 4\n")

#Summaries and plots for model 4
summary(fit_gershman4_vb)
```


(VB) Run model 5

```{r}
startTime = Sys.time(); print(startTime) 

fit_gershman5_vb<- vb(gershman_model5,
                data    = GershmandataList,
                seed    = 1450154626) 

endTime = Sys.time(); print(endTime) 
cat("It took",as.character.Date(endTime - startTime), "to fit Gershman Model 5\n")

#Summaries and plots for model 5
summary(fit_gershman5_vb)
```


(MCMC) Run model 1

```{r}
nIter<-2000  #how many samples will be taken
nChains<- 4    #how many starting points of the sampler (robot) (sequences of visits)
nWarmup<- floor(nIter/2) #how many of the samples will be discarded as warmups
nThin<- 1   #how many samples will be included/skipped

cat("Estimating", gershman_model1file, "model... \n")
    startTime = Sys.time(); print(startTime)
    cat("Calling", nChains, "simulations in Stan... \n")

#Use this to see how long the modelling takes and each simulation
fit_gershman1_MCMC <- stan(gershman_model1file,
                   data    = GershmandataList,
                   chains  = nChains,
                   iter    = nIter,
                   warmup  = nWarmup,
                   thin    = nThin,
                   init    = "random",
                   seed    = 1450154626) #so you can replicate the model

# save.image(file = "Gershman_MCMC_model1.RData")

cat("Finishing", gershman_model1file, "model simulation ... \n")
    endTime = Sys.time(); print(endTime)
    cat("It took",as.character.Date(endTime - startTime), "\n")

# Model Summary and Diagnostics, use this to see how long the modelling takes and each simulation
 print(fit_gershman1_MCMC) #look at the model output

 summary(fit_gershman1_MCMC)
```


(MCMC) Run model 2

```{r}
nIter<-100  #how many samples will be taken
nChains<- 1    #how many starting points of the sampler (robot) (sequences of visits)
nWarmup<- floor(nIter/2) #how many of the samples will be discarded as warmups
nThin<- 1   #how many samples will be included/skipped

cat("Estimating", gershman_model2file, "model... \n")
    startTime = Sys.time(); print(startTime)
    cat("Calling", nChains, "simulations in Stan... \n")

#Use this to see how long the modelling takes and each simulation
fit_gershman2_MCMC <- stan(gershman_model2file,
                   data    = GershmandataList,
                   chains  = nChains,
                   iter    = nIter,
                   warmup  = nWarmup,
                   thin    = nThin,
                   init    = "random",
                   seed    = 1450154626) #so you can replicate the model

# save.image(file = "Gershman_MCMC_model2.RData")

cat("Finishing", gershman_model2file, "model simulation ... \n")
    endTime = Sys.time(); print(endTime)
    cat("It took",as.character.Date(endTime - startTime), "\n")

# Model Summary and Diagnostics, use this to see how long the modelling takes and each simulation
 print(fit_gershman2_MCMC) #look at the model output

 summary(fit_gershman2_MCMC)
```


(MCMC) Run model 3

```{r}
nIter<-100  #how many samples will be taken
nChains<- 1    #how many starting points of the sampler (robot) (sequences of visits)
nWarmup<- floor(nIter/2) #how many of the samples will be discarded as warmups
nThin<- 1   #how many samples will be included/skipped

cat("Estimating", gershman_model3file, "model... \n")
    startTime = Sys.time(); print(startTime)
    cat("Calling", nChains, "simulations in Stan... \n")

#Use this to see how long the modelling takes and each simulation
fit_gershman3_MCMC <- stan(gershman_model3file,
                   data    = GershmandataList,
                   chains  = nChains,
                   iter    = nIter,
                   warmup  = nWarmup,
                   thin    = nThin,
                   init    = "random",
                   seed    = 1450154626) #so you can replicate the model

# save.image(file = "Gershman_MCMC_model3.RData")

cat("Finishing", gershman_model3file, "model simulation ... \n")
    endTime = Sys.time(); print(endTime)
    cat("It took",as.character.Date(endTime - startTime), "\n")

# Model Summary and Diagnostics, use this to see how long the modelling takes and each simulation
 print(fit_gershman3_MCMC) #look at the model output

 summary(fit_gershman3_MCMC)
```


(MCMC) Run model 4

```{r}
nIter<-100  #how many samples will be taken
nChains<- 1    #how many starting points of the sampler (robot) (sequences of visits)
nWarmup<- floor(nIter/2) #how many of the samples will be discarded as warmups
nThin<- 1   #how many samples will be included/skipped

cat("Estimating", gershman_model4file, "model... \n")
    startTime = Sys.time(); print(startTime)
    cat("Calling", nChains, "simulations in Stan... \n")

#Use this to see how long the modelling takes and each simulation
fit_gershman4_MCMC <- stan(gershman_model4file,
                   data    = GershmandataList,
                   chains  = nChains,
                   iter    = nIter,
                   warmup  = nWarmup,
                   thin    = nThin,
                   init    = "random",
                   seed    = 1450154626) #so you can replicate the model

# save.image(file = "Gershman_MCMC_model4.RData")

cat("Finishing", gershman_model4file, "model simulation ... \n")
    endTime = Sys.time(); print(endTime)
    cat("It took",as.character.Date(endTime - startTime), "\n")

# Model Summary and Diagnostics, use this to see how long the modelling takes and each simulation
 print(fit_gershman4_MCMC) #look at the model output

 summary(fit_gershman4_MCMC)
```


(MCMC) Run model 5

```{r}
nIter<-100  #how many samples will be taken
nChains<- 1    #how many starting points of the sampler (robot) (sequences of visits)
nWarmup<- floor(nIter/2) #how many of the samples will be discarded as warmups
nThin<- 1   #how many samples will be included/skipped

cat("Estimating", gershman_model5file, "model... \n")
    startTime = Sys.time(); print(startTime)
    cat("Calling", nChains, "simulations in Stan... \n")

#Use this to see how long the modelling takes and each simulation
fit_gershman5_MCMC <- stan(gershman_model5file,
                   data    = GershmandataList,
                   chains  = nChains,
                   iter    = nIter,
                   warmup  = nWarmup,
                   thin    = nThin,
                   init    = "random",
                   seed    = 1450154626) #so you can replicate the model

# save.image(file = "Gershman_MCMC_model5.RData")

cat("Finishing", gershman_model5file, "model simulation ... \n")
    endTime = Sys.time(); print(endTime)
    cat("It took",as.character.Date(endTime - startTime), "\n")

# Model Summary and Diagnostics, use this to see how long the modelling takes and each simulation
 print(fit_gershman5_MCMC) #look at the model output

 summary(fit_gershman5_MCMC)
```
